[feat/#128] - 처음에 모든 메세지들을 다 가져오기 보다는 조금만 가져와서 호출 줄이기
