+++ src/components/chat/chatBody/ChatList.tsx
@@ -24,12 +24,7 @@ const ChatList = ({ user, chatRoomId }: { user: User | null; chatRoomId: string 
   const [count, setCount] = useState(1);
   const room = useRoomDataQuery(chatRoomId);
   const roomId = room?.roomId;
-
-  // const results = useRecruitingMyroomQuery(user?.id as string);
-  // console.log(results);
-
-  // const roomData = useRoomInfoWithRoomIdQuery(roomId as UUID);
-  // console.log(roomData);
+  console.log(hasMore);
 
   useEffect(() => {
     if (roomId && chatRoomId) {
@@ -66,14 +61,16 @@ const ChatList = ({ user, chatRoomId }: { user: User | null; chatRoomId: string 
     }
   }, [messages, setMessages, isScrolling, roomId, chatRoomId]);
 
+  // 처음에 로드될 시
   useEffect(() => {
-    // 스크롤 중이 아니면 처음에 로드될 시 기본적으로 스크롤 다운
+    // 스크롤 중이 아니면 기본적으로 스크롤 다운
     const scrollBox = scrollRef.current;
     if (scrollBox && isScrolling === false) {
       scrollBox.scrollTop = scrollBox.scrollHeight;
     }
   }, [messages, isScrolling]);
 
+  // 스크롤 이벤트가 발생할 때
   const handleScroll = () => {
     const scrollBox = scrollRef.current;
     if (scrollBox) {
@@ -82,6 +79,7 @@ const ChatList = ({ user, chatRoomId }: { user: User | null; chatRoomId: string 
       if (!isScroll) {
         setNewAddedMsgNum(0);
       }
+      // console.log('여기 scrollTop', scrollBox.scrollTop);
       setIsScrollTop(scrollBox.scrollTop === 0);
     }
   };
@@ -90,29 +88,6 @@ const ChatList = ({ user, chatRoomId }: { user: User | null; chatRoomId: string 
     scrollRef.current.scrollTop = scrollRef.current.scrollHeight;
   };
   // insert 할 때 없어졌으면 좋겠는데..
-
-  const fetchMoreMsg = async () => {
-    const { from, to } = getFromTo(count, ITEM_INTERVAL);
-    const { error, data: newMsgs } = await clientSupabase
-      .from('messages')
-      .select('*')
-      .range(from, to)
-      .eq('chatting_room_id', String(chatRoomId))
-      .order('created_at', { ascending: false });
-    if (error) {
-      alert('이전 메세지를 불러오는 데에 오류가 발생했습니다.');
-    } else {
-      setMessages([...(newMsgs ? newMsgs.reverse() : []), ...messages]);
-      if (newMsgs.length < ITEM_INTERVAL + 1) {
-        setHasMore(false);
-      } else {
-        setCount((prev) => prev + 1);
-      }
-    }
-  };
-  // 더보기를 누르면 다시 렌더링이 되면서 useEffect가 실행되어 scrollTop이랑 scrollHeight가 같아져야 하는데(스크롤다운) 왜 스크롤이 안내려가지는지?
-  // 더보기 눌렀을 때 위치 다시 생각해봐야함
-  // 삭제 후 더보기 누르면 제대로 안 불러와짐
 
   return (
     <>
@@ -123,7 +98,7 @@ const ChatList = ({ user, chatRoomId }: { user: User | null; chatRoomId: string 
       >
         <ChatSearch isScrollTop={isScrollTop} />
 
-        {hasMore && <LoadChatMore fetchMoreMsg={fetchMoreMsg} />}
+        {hasMore ? <LoadChatMore chatRoomId={chatRoomId} count={count} setCount={setCount} /> : <></>}
         {messages?.map((msg) => {
           if (msg.send_from === user?.id) {
             return <MyChat msg={msg} key={msg.message_id} />;
