+++ src/components/chat/chatBody/ChatSearch.tsx
@@ -9,6 +9,7 @@ const ChatSearch = ({ isScrollTop }: { isScrollTop: boolean }) => {
   const [doneSearchDivs, setDoneSearchdivs] = useState<(HTMLElement | null)[]>();
   const [searchCount, setSearchCount] = useState(0);
   const [upDownCount, setUpDownCount] = useState(0);
+  // console.log('isScrollTop', isScrollTop);
 
   const handleSearch = () => {
     if (searchWord) {
